requirements.txt
errbot==5.1.3
jsonpickle==0.9.6
python-dotenv==0.8.2
redis==2.10.6
slackclient==1.1.3
